+++ src/main/java/com/andersmmg/falloutstuff/block/custom/ToggleButtonBlock.java
@@ -1,0 +1,121 @@
+package com.andersmmg.falloutstuff.block.custom;
+
+import com.andersmmg.falloutstuff.sound.ModSounds;
+import com.andersmmg.falloutstuff.util.VoxelUtils;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockRenderType;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemPlacementContext;
+import net.minecraft.sound.SoundCategory;
+import net.minecraft.state.StateManager;
+import net.minecraft.state.property.BooleanProperty;
+import net.minecraft.state.property.DirectionProperty;
+import net.minecraft.state.property.Properties;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.BlockMirror;
+import net.minecraft.util.BlockRotation;
+import net.minecraft.util.Hand;
+import net.minecraft.util.hit.BlockHitResult;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Direction;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.world.BlockView;
+import net.minecraft.world.World;
+import net.minecraft.world.event.GameEvent;
+
+public class ToggleButtonBlock extends Block {
+    public static final DirectionProperty FACING = Properties.HORIZONTAL_FACING;
+    public static final BooleanProperty POWERED = Properties.POWERED;
+    private static final VoxelShape VOXEL_SHAPE = Block.createCuboidShape(5, 4, 14, 11, 12, 16);
+
+    public ToggleButtonBlock(Settings settings) {
+        super(settings);
+        this.setDefaultState(this.stateManager.getDefaultState().with(FACING, Direction.NORTH).with(POWERED, false));
+    }
+
+    protected static Direction getDirection(BlockState state) {
+        return state.get(FACING);
+    }
+
+    @Override
+    public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
+        BlockState blockState;
+        if (world.isClient) {
+            blockState = state.cycle(POWERED);
+            world.setBlockState(pos, blockState, 3);
+            return ActionResult.SUCCESS;
+        } else {
+            blockState = this.togglePower(state, world, pos);
+            float f = blockState.get(POWERED) ? 0.6F : 0.5F;
+            world.playSound(null, pos, ModSounds.TOGGLE_CLICK, SoundCategory.BLOCKS, 0.9F, f);
+            world.emitGameEvent(player, blockState.get(POWERED) ? GameEvent.BLOCK_ACTIVATE : GameEvent.BLOCK_DEACTIVATE, pos);
+            return ActionResult.CONSUME;
+        }
+    }
+
+    public BlockState togglePower(BlockState state, World world, BlockPos pos) {
+        state = state.cycle(POWERED);
+        world.setBlockState(pos, state, 3);
+        BlockState adjacentState = world.getBlockState(pos.offset(getDirection(state).getOpposite(), 2));
+        if (adjacentState.isOf(this) && adjacentState.get(FACING) == getDirection(state).getOpposite()) {
+            world.setBlockState(pos.offset(getDirection(state).getOpposite(), 2), adjacentState.with(POWERED, state.get(POWERED)), 3);
+        }
+        this.updateNeighbors(state, world, pos);
+        return state;
+    }
+
+    private void updateNeighbors(BlockState state, World world, BlockPos pos) {
+        world.updateNeighborsAlways(pos, this);
+        world.updateNeighborsAlways(pos.offset(getDirection(state).getOpposite()), this);
+    }
+
+    @Override
+    public BlockRenderType getRenderType(BlockState state) {
+        return BlockRenderType.MODEL;
+    }
+
+    @Override
+    public BlockState rotate(BlockState state, BlockRotation rotation) {
+        return state.with(FACING, rotation.rotate(state.get(FACING)));
+    }
+
+    @Override
+    public BlockState mirror(BlockState state, BlockMirror mirror) {
+        return state.rotate(mirror.getRotation(state.get(FACING)));
+    }
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return VoxelUtils.rotateShape(getDirection(state), VOXEL_SHAPE);
+    }
+
+    @Override
+    protected void appendProperties(StateManager.Builder<Block, BlockState> builder) {
+        builder.add(FACING, POWERED);
+    }
+
+    @Override
+    public BlockState getPlacementState(ItemPlacementContext ctx) {
+        if (ctx.getSide().getAxis().isHorizontal()) {
+            return this.getDefaultState().with(FACING, ctx.getSide());
+        }
+        return null;
+    }
+
+    @Override
+    public boolean emitsRedstonePower(BlockState state) {
+        return true;
+    }
+
+    @Override
+    public int getStrongRedstonePower(BlockState state, BlockView world, BlockPos pos, Direction direction) {
+        return state.get(POWERED) ? 15 : 0;
+    }
+
+    @Override
+    public int getWeakRedstonePower(BlockState state, BlockView world, BlockPos pos, Direction direction) {
+        return state.get(POWERED) ? 15 : 0;
+    }
+}